Remove commute and replace it with travel allowance (something the user can set in the settings page). New fee for weekend and holidays. New pricing engine, lessons stores all to be independent.

+++ Apolo/ViewModels/InvoicesViewModel.cs
@@ -203,7 +203,7 @@ namespace Apolo.ViewModels
 
             var list = attendances.ToArray();
             decimal subTotal = list.Sum(a => a.Price);
-            decimal ivaPercent = user.IvaPercent;
+            decimal ivaPercent = (decimal)user.IvaPercent;
             decimal ivaAmount = Math.Round(subTotal * ivaPercent / 100m, 2, MidpointRounding.AwayFromZero);
             var total = subTotal + ivaAmount;
